pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "gmail-reader"
version = "0.1.0"
description = "Read-only Gmail investigation tool"
requires-python = ">=3.9"
dependencies = [
    "google-api-python-client>=2.100.0",
    "google-auth-httplib2>=0.1.1",
    "google-auth-oauthlib>=1.1.0",
    "python-dotenv>=1.0.0",
    "tabulate>=0.9.0",
    "mcp>=1.0.0",
]

[project.optional-dependencies]
test = ["pytest>=7.0.0"]

[project.scripts]
gmail-reader = "gmail_reader.__main__:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
